requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-embed-video==1.4.0
django-extensions==3.1.3
django-simplecms==1.0.5
django-tidy==0.1
easy-thumbnails==2.7.1
idna==2.10
libsass==0.21.0
Markdown==3.3.4
Pillow==8.3.0
psycopg2==2.9.1
pylibmc==1.6.1
pytidylib==0.3.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
